requirements.txt
bitstring
numpy
pyserial
scipy
smbus
RPi.GPIO
Adafruit_DHT